No issue: added smoke tests from MediaNotificationTest.kt
